requirements.txt
pycurl==7.19.0
pyzmq==13.1.0
wsgiref==0.1.2
